fix parse server init freeze
