+++ src/functions/schema/schema.ts
@@ -201,8 +201,9 @@ const getAllParseSchema = async () => {
   try {
     return await Parse.Schema.all();
   } catch (e) {
+    await ((global as Record<string, any>).reconfigureServer as Function)();
     await wait(1000);
-    return getAllParseSchema();
+    return await getAllParseSchema();
   }
 };
